+++ apps/mobile_flutter/lib/enums/lookup_status.dart
