+++ apps/mobile_flutter/lib/enums/refresh_status.dart
@@ -1,0 +1,3 @@
+enum RefreshStatus {
+  REFRESHING,
+}
