+++ apps/mobile_flutter/lib/views/settings/settings_utils.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_weather/enums/enums.dart';
 import 'package:flutter_weather/localization.dart';
-import 'package:flutter_weather/views/settings/settings_enums.dart';
 
 UpdatePeriod? getPeriod({
   BuildContext? context,
